requirements.txt
google-genai
jinja2
pydantic
pytest
aiohttp
